# Make engine routes filterable in bin/rails routes, improve engine formatting