--- a/actionpack/lib/action_dispatch/routing/inspector.rb
+++ b/actionpack/lib/action_dispatch/routing/inspector.rb
@@ -64,6 +64,14 @@ module ActionDispatch
       def engine?
         app.engine?
       end
+
+      def to_h
+        { name: name,
+          verb: verb,
+          path: path,
+          reqs: reqs,
+          source_location: source_location }
+      end
     end
 
     ##
@@ -72,33 +80,51 @@ module ActionDispatch
     # not use this class.
     class RoutesInspector # :nodoc:
       def initialize(routes)
-        @engines = {}
-        @routes = routes
+        @routes = wrap_routes(routes)
+        @engines = load_engines_routes
       end
 
       def format(formatter, filter = {})
-        routes_to_display = filter_routes(normalize_filter(filter))
-        routes = collect_routes(routes_to_display)
-        if routes.none?
-          formatter.no_routes(collect_routes(@routes), filter)
-          return formatter.result
-        end
+        all_routes = { nil => @routes }.merge(@engines)
 
-        formatter.header routes
-        formatter.section routes
-
-        @engines.each do |name, engine_routes|
-          formatter.section_title "Routes for #{name}"
-          if engine_routes.any?
-            formatter.header engine_routes
-            formatter.section engine_routes
-          end
+        all_routes.each do |engine_name, routes|
+          format_routes(formatter, filter, engine_name, routes)
         end
 
         formatter.result
       end
 
       private
+        def format_routes(formatter, filter, engine_name, routes)
+          routes = filter_routes(routes, normalize_filter(filter)).map(&:to_h)
+
+          formatter.section_title "Routes for #{engine_name || "application"}" if @engines.any?
+          if routes.any?
+            formatter.header routes
+            formatter.section routes
+          else
+            formatter.no_routes engine_name, routes, filter
+          end
+          formatter.footer routes
+        end
+
+        def wrap_routes(routes)
+          routes.routes.map { |route| RouteWrapper.new(route) }.reject(&:internal?)
+        end
+
+        def load_engines_routes
+          engine_routes = @routes.select(&:engine?)
+
+          engines = engine_routes.to_h do |engine_route|
+            engine_app_routes = engine_route.rack_app.routes
+            engine_app_routes = engine_app_routes.routes if engine_app_routes.is_a?(ActionDispatch::Routing::RouteSet)
+
+            [engine_route.endpoint, wrap_routes(engine_app_routes)]
+          end
+
+          engines
+        end
+
         def normalize_filter(filter)
           if filter[:controller]
             { controller: /#{filter[:controller].underscore.sub(/_?controller\z/, "")}/ }
@@ -118,39 +144,13 @@ module ActionDispatch
           end
         end
 
-        def filter_routes(filter)
+        def filter_routes(routes, filter)
           if filter
-            @routes.select do |route|
-              route_wrapper = RouteWrapper.new(route)
-              filter.any? { |filter_type, value| route_wrapper.matches_filter?(filter_type, value) }
+            routes.select do |route|
+              filter.any? { |filter_type, value| route.matches_filter?(filter_type, value) }
             end
           else
-            @routes
-          end
-        end
-
-        def collect_routes(routes)
-          routes.collect do |route|
-            RouteWrapper.new(route)
-          end.reject(&:internal?).collect do |route|
-            collect_engine_routes(route)
-
-            { name: route.name,
-              verb: route.verb,
-              path: route.path,
-              reqs: route.reqs,
-              source_location: route.source_location }
-          end
-        end
-
-        def collect_engine_routes(route)
-          name = route.endpoint
-          return unless route.engine?
-          return if @engines[name]
-
-          routes = route.rack_app.routes
-          if routes.is_a?(ActionDispatch::Routing::RouteSet)
-            @engines[name] = collect_routes(routes.routes)
+            routes
           end
         end
     end
@@ -174,27 +174,36 @@ module ActionDispatch
         def header(routes)
         end
 
-        def no_routes(routes, filter)
-          @buffer <<
-            if routes.none?
-              <<~MESSAGE
-                You don't have any routes defined!
+        def footer(routes)
+        end
 
-                Please add some routes in config/routes.rb.
-              MESSAGE
-            elsif filter.key?(:controller)
+        def no_routes(engine, routes, filter)
+          @buffer <<
+            if filter.key?(:controller)
               "No routes were found for this controller."
             elsif filter.key?(:grep)
               "No routes were found for this grep pattern."
+            elsif routes.none?
+              if engine
+                "No routes defined."
+              else
+                <<~MESSAGE
+                  You don't have any routes defined!
+
+                  Please add some routes in config/routes.rb.
+                MESSAGE
+              end
             end
 
-          @buffer << "For more information about routes, see the Rails guide: https://guides.rubyonrails.org/routing.html."
+          unless engine
+            @buffer << "For more information about routes, see the Rails guide: https://guides.rubyonrails.org/routing.html."
+          end
         end
       end
 
       class Sheet < Base
         def section_title(title)
-          @buffer << "\n#{title}:"
+          @buffer << "#{title}:"
         end
 
         def section(routes)
@@ -203,6 +212,10 @@ module ActionDispatch
 
         def header(routes)
           @buffer << draw_header(routes)
+        end
+
+        def footer(routes)
+          @buffer << ""
         end
 
         private
@@ -235,11 +248,15 @@ module ActionDispatch
         end
 
         def section_title(title)
-          @buffer << "\n#{"[ #{title} ]"}"
+          @buffer << "#{"[ #{title} ]"}"
         end
 
         def section(routes)
           @buffer << draw_expanded_section(routes)
+        end
+
+        def footer(routes)
+          @buffer << ""
         end
 
         private
@@ -272,7 +289,7 @@ module ActionDispatch
           super
         end
 
-        def no_routes(routes, filter)
+        def no_routes(engine, routes, filter)
           @buffer <<
             if filter.none?
               "No unused routes found."
@@ -301,6 +318,9 @@ module ActionDispatch
 
       # The header is part of the HTML page, so we don't construct it here.
       def header(routes)
+      end
+
+      def footer(routes)
       end
 
       def no_routes(*)