### Motivation / Background

This Pull Request has been created because as mentioned in https://github.com/rails/rails/pull/54573, the output of `bin/rails routes` doesn't filter engine routes properly.

Closes https://github.com/rails/rails/pull/54573

### Detail

This Pull Request changes the routing inspect to do the following:
- Always unfurl engine routes so that all routes of an application can be filtered as needed
- Improve heading, messaging, and formatting of the inspector to denote app routes, when engines have no routes, spacing between engines, etc.
- Only adds additional formatting on applications using engines

### Additional information

Since we rewrote how routes are collected and printed, I did a quick `time` benchmark on a large app to make sure we're not slowing things down:

Before:
```
bin/rails routes  4.73s user 3.26s system 84% cpu 9.513 total
bin/rails routes -g filter  4.75s user 3.14s system 90% cpu 8.766 total
```

After:
```
bin/rails routes  4.44s user 2.68s system 75% cpu 9.431 total
bin/rails routes -g filter  4.88s user 3.19s system 90% cpu 8.889 total
```

Same-ish to a bit faster on a full print. Applying a filter is slightly slower, but that's expected because it iterates through all routes (app and engine) now instead of lazy unfurling that previously missed embedded routes.

### Checklist

Before submitting the PR make sure the following are checked:

* [x] This Pull Request is related to one change. Unrelated changes should be opened in separate PRs.
* [x] Commit message has a detailed description of what changed and why. If this PR fixes a related issue include it in the commit message. Ex: `[Fix #issue-number]`
* [x] Tests are added or updated if you fix a bug or add a feature.
* [x] CHANGELOG files are updated for the changed libraries if there is a behavior change or additional feature. Minor bug fixes and documentation changes should not be included.